[realm] some more composition + renaming
