+++ wow/core/src/main/scala/wow/realm/session/NetworkWorker.scala
@@ -3,55 +3,35 @@ package wow.realm.session
 import java.util.concurrent.ThreadLocalRandom
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Props}
-import scodec.Codec
 import scodec.bits.BitVector
-import wow.common.network.{EventIncoming, TCPSession, TCPSessionFactory}
-import wow.realm.crypto.SessionCipher
-import wow.realm.handlers.HandledBy
+import wow.common.network.{TCPSession, TCPSessionFactory}
 import wow.realm.protocol.payloads.ServerAuthChallenge
 import wow.realm.protocol.{OpCodes, _}
-import wow.realm.session.NetworkWorker.HandlePacket
 import wow.realm.{RealmContext, RealmContextData}
 
-import scala.concurrent.duration._
 import scala.language.postfixOps
 import scala.util.Random
 
 /**
   * Handles a realm session's networking
+  * Most of the features of this actor are actual provided by composing in traits
   */
 class NetworkWorker(connection: ActorRef)(override implicit val realm: RealmContextData)
   extends TCPSession(connection)
-          with BufferIncomingPackets
+          with HandleIncomingPackets
+          with SendForwardedPackets
+          with SendPacketsToClient
           with Actor
           with ActorLogging
-          with PacketHandlerTag
-          with RealmContext
-          with CanSendPackets {
-  private[session] var sessionCipher: Option[SessionCipher] = None
-
+          with RealmContext {
   // Send initial challenge packet
   sendAuthChallenge()
 
-  override def receive: Receive = super[TCPSession].receive orElse super[BufferIncomingPackets].receive orElse {
-    case NetworkWorker.SendRawPayload(payloadBits, opCode) =>
-      sendRaw(payloadBits, opCode)
-
-    case NetworkWorker.SendRaw(bits) =>
-      sendRaw(bits)
-
-    case NetworkWorker.SendRawSplit(headerBits, payloadBits) =>
-      val bits = PacketSerialization.outgoing(headerBits, payloadBits)(sessionCipher)
-
-      sendRaw(bits)
-
-    case NetworkWorker.Terminate(delayed) =>
-      if (delayed) {
-        terminateDelayed()
-      } else {
-        terminateNow()
-      }
-  }
+  override def receive: Receive = Seq(
+    super[TCPSession].receive,
+    super[HandleIncomingPackets].receive,
+    super[SendForwardedPackets].receive
+  ).reduceLeft(_.andThen(_))
 
   /**
     * This must remain lazy or be moved before the call to sendAuthChallenge().
@@ -76,50 +56,6 @@ class NetworkWorker(connection: ActorRef)(override implicit val realm: RealmCont
     )
 
     sendPayload(authChallenge)
-  }
-
-  private val terminationDelay = 1 second
-
-  override def terminateDelayed(): Unit = {
-    context.system.scheduler.scheduleOnce(terminationDelay)(terminateNow())(context.dispatcher)
-  }
-
-  override def terminateNow(): Unit = {
-    disconnect()
-    context.stop(self)
-  }
-
-  override def sendPayload[A <: Payload with ServerSide](payload: A)
-    (implicit codec: Codec[A], opCodeProvider: OpCodeProvider[A]): Unit = {
-    log.debug("Sending " + payload)
-    val bits = PacketSerialization.outgoing(payload)(sessionCipher)
-
-    sendRaw(bits)
-  }
-
-  override def sendRaw(payloadBits: BitVector, opCode: OpCodes.Value): Unit = {
-    val bits = PacketSerialization.outgoing(payloadBits, opCode)(sessionCipher)
-
-    sendRaw(bits)
-  }
-
-  override def sendRaw(headerBits: BitVector, payloadBits: BitVector): Unit = {
-    val bits = PacketSerialization.outgoing(headerBits, payloadBits)(sessionCipher)
-
-    sendRaw(bits)
-  }
-
-  override def sendRaw(bits: BitVector): Unit = outgoing(bits)
-
-  /**
-    * Enables the cipher for all packets onwards
-    *
-    * @param sessionKey session key to use
-    */
-  def enableCipher(sessionKey: BigInt): Unit = {
-    require(sessionCipher.isEmpty)
-    log.debug(s"Session key set up: $sessionKey")
-    sessionCipher = Some(new SessionCipher(sessionKey))
   }
 
   /**
@@ -204,7 +140,6 @@ object NetworkWorker {
   case class Terminate(delayed: Boolean)
 
   case class HandlePacket(header: ClientHeader, payloadBits: BitVector)
-
 }
 
 class NetworkWorkerFactory(implicit realm: RealmContextData) extends TCPSessionFactory {
